+++ webpack.prod.config.js
@@ -31,7 +31,8 @@ var config = {
       commonjs2: 'react',
       commonjs: 'react',
       amd: 'react'
-    }
+    },
+    'react-native': true,
   },
 };